Struct improv::Bounds

pub struct Bounds {
    // some fields omitted
}

A rectangular bounding box with an x and y origin, and a width and height.

Methods

impl Bounds

fn new(width: u32, height: u32) -> Bounds

Create a new Bounds from width and height and an origin of (0,0)

fn aspect_ratio(&self) -> f32

w / h as a fractional number

fn area(&self) -> usize

The total number of pixels that lie within the bounds

fn width(&self) -> u32

fn height(&self) -> u32

fn w(&self) -> u32

A convenient method for width()

fn h(&self) -> u32

A convenient alias for height()

fn origin(&self) -> Point